+++ semana15/projeto-sistema-bancario/src/controllers/BalanceController.ts
@@ -1,6 +1,5 @@
-import dayjs from "dayjs";
 import { Request, Response } from "express";
-import { Extract, User } from "../data/users";
+import { User } from "../data/users";
 import { verifyCpf } from "../utilities/verifiers";
 
 class BalanceController {
@@ -21,30 +20,6 @@ class BalanceController {
       res.status(200).send({ balance: result });
     } catch (error) {
       res.status(errorCode).send({ message: error.message });
-    }
-  }
-
-  async execute(req: Request, res: Response) {
-    let errorCode: number = 400;
-    try {
-      if (!req.body.name || !req.body.cpf || !req.body.value) {
-        errorCode = 422;
-        throw new Error("Preenchimento inválido, tente novamente!");
-      }
-      const accountExist: User | undefined = verifyCpf(req.body.cpf);
-      if (!accountExist) {
-        errorCode = 404;
-        throw new Error("Conta não existe em nosso banco de dados!");
-      }
-      const newDeposit: Extract = {
-        value: Number(req.body.value),
-        date: dayjs().format("DD/MM/YYYY"),
-        description: "Depósito de dinheiro",
-      };
-      accountExist.information.push(newDeposit);
-      res.status(202).send("Depósito realizado com sucesso!");
-    } catch (error) {
-      res.status(errorCode).send(error.message);
     }
   }
 }
